refactor(UIKIT-417,validations): Flow переименован в compose
