+++ packages/validations/src/composeAllSettled/composeAllSettled.test.ts
@@ -1,8 +1,8 @@
-import { flowAll } from './flowAll';
+import { composeAllSettled } from './composeAllSettled';
 
-describe('flowAll', () => {
+describe('composeAllSettled', () => {
   it('Возвращает все ошибки, которые появляется при выполнении правил', () => {
-    const validate = flowAll(
+    const validate = composeAllSettled(
       () => 'error1',
       () => 'error2',
     );
@@ -11,7 +11,7 @@ describe('flowAll', () => {
   });
 
   it('Делает многомерные массивы ошибок плоскими', () => {
-    const validate = flowAll(
+    const validate = composeAllSettled(
       () => ['error1', 'error2'],
       () => 'error3',
     );
@@ -20,7 +20,7 @@ describe('flowAll', () => {
   });
 
   it('Не добавляет в результирующий массив ошибок undefined', () => {
-    const validate = flowAll(
+    const validate = composeAllSettled(
       () => ['error1', 'error2'],
       () => undefined,
       () => 'error3',
@@ -30,7 +30,7 @@ describe('flowAll', () => {
   });
 
   it('Возвращает undefined, если нет ошибок', () => {
-    const validate = flowAll(
+    const validate = composeAllSettled(
       () => undefined,
       () => undefined,
     );
@@ -39,7 +39,7 @@ describe('flowAll', () => {
   });
 
   it('Если итоговая ошибка одна, то она возвращается без массива', () => {
-    const validate = flowAll(
+    const validate = composeAllSettled(
       () => undefined,
       () => 'error',
     );
@@ -48,9 +48,9 @@ describe('flowAll', () => {
   });
 
   it('Поддерживает вложенность', () => {
-    const validate = flowAll(
+    const validate = composeAllSettled(
       () => ['error1', 'error2'],
-      flowAll(() => ['error3', 'error4']),
+      composeAllSettled(() => ['error3', 'error4']),
     );
 
     expect(validate(null)).toEqual(['error1', 'error2', 'error3', 'error4']);
