+++ packages/validations/src/composeToString/composeToString.ts
@@ -1,16 +1,16 @@
 import { InitializedRule } from '../types';
-import { flow } from '../flow';
+import { compose } from '../compose';
 
 /**
  * @description Объеденяет правила в цепочку и форматирует результат в одну строку. Останаваливается при возникновении ошибки. Работает слева направо
  * Подходит в качестве адаптера к react-final-form
- * @example flowToString(isIncludeDot(), isIncludeComma());
- * @example flowToString(isRequired(), flowToString(isIncludeDot(), isIncludeComma()));
+ * @example composeToString(isIncludeDot(), isIncludeComma());
+ * @example composeToString(isRequired(), composeToString(isIncludeDot(), isIncludeComma()));
  */
-export const flowToString =
+export const composeToString =
   (...rules: InitializedRule[]) =>
   (value?: unknown): string | undefined => {
-    const validationResult = flow(...rules)(value);
+    const validationResult = compose(...rules)(value);
 
     if (Array.isArray(validationResult)) {
       return validationResult[0];
