--- packages/validations/src/flowAll/index.ts
@@ -1,1 +1,0 @@
-export * from './flowAll';
